// packages/webdev-api-contracts/src/supabase.schema.ts
export const publicQuestionRowSchema = z.object({
  answer: z.number(),
  category: z.string().nullable(),
  created_at: z.string(),
  creator_id: z.string(),
  explanation: z.string().nullable(),
  id: z.string(),
  option_a: z.string(),
  option_b: z.string(),
  option_c: z.string().nullable(),
  option_d: z.string().nullable(),
  question: z.string(),
  schedule: z.string(),
  value: z.number(),
});

export const publicQuestionInsertSchema = z.object({
  answer: z.number(),
  category: z.string().optional().nullable(),
  created_at: z.string().optional(),
  creator_id: z.string(),
  explanation: z.string().optional().nullable(),
  id: z.string().optional(),
  option_a: z.string(),
  option_b: z.string(),
  option_c: z.string().optional().nullable(),
  option_d: z.string().optional().nullable(),
  question: z.string(),
  schedule: z.string(),
  value: z.number(),
});